Guessing game -- initial
We code up a simple guessing game directly in the main function.
This is obviously a bad design, but we start with the obvious and
then we will refactor it multiple times.

The code is mirroring the initial example from the official Rust docs
https://doc.rust-lang.org/book/ch02-00-guessing-game-tutorial.html

We added a dependency using `cargo add rand` using the utility
cargo-edit.

There is a whole lot of Enums used in the body, we recommend reading
up on them.